+++ CHANGELOG.md
@@ -6,7 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [1.1.0] - 2015-05-11
 ### Added
-- Remove exception if Response in not OK
+- Remove exception if Response is not OK
 - Five more tests to Response/DecryptResponse
 - Missing PHPdoc